+++ socialhome/federate/tasks.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timezone, timedelta
 import logging
 from typing import List, TYPE_CHECKING, Optional
 from uuid import uuid4
@@ -6,15 +7,18 @@ import django_rq
 from django.conf import settings
 from dynamic_preferences.registries import global_preferences_registry
 from federation.entities import base
+from federation.entities.activitypub.constants import NAMESPACE_PUBLIC
+from federation.entities.activitypub.enums import ActivityType
 from federation.exceptions import NoSuitableProtocolFoundError, NoSenderKeyFoundError, SignatureVerificationError
 from federation.inbound import handle_receive
 from federation.outbound import handle_send
 
+from socialhome.activities.models import Activity
 from socialhome.content.enums import ContentType
 from socialhome.content.models import Content
 from socialhome.enums import Visibility
 from socialhome.federate.models import Payload
-from socialhome.federate.utils.tasks import process_entities, sender_key_fetcher
+from socialhome.federate.utils.tasks import process_entities, sender_key_fetcher, process_replies
 from socialhome.federate.utils import make_federable_profile, get_outbound_payload_logger
 from socialhome.federate.utils.entities import make_federable_content, make_federable_retraction
 from socialhome.users.models import Profile
@@ -92,7 +96,7 @@ def send_content(content_id, activity_fid, recipient_id=None):
     entity = make_federable_content(content)
     if entity:
         entity.activity_id = activity_fid
-        if settings.DEBUG:
+        if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
             # Don't send in development mode
             return
         recipients = []
@@ -171,7 +175,7 @@ def send_reply(content_id, activity_fid):
     if not entity:
         logger.warning("send_reply - No entity for %s", content)
     entity.activity_id = activity_fid
-    if settings.DEBUG:
+    if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
         # Don't send in development mode
         return
     recipients = []
@@ -210,8 +214,8 @@ def send_share(content_id, activity_fid):
         return
     entity = make_federable_content(content)
     if entity:
-        entity.activity_id = activity_fid
-        if settings.DEBUG:
+        entity.id = activity_fid
+        if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
             # Don't send in development mode
             return
         recipients = _get_remote_followers(content.author, content.visibility)
@@ -220,6 +224,15 @@ def send_share(content_id, activity_fid):
             recipients.append(content.share_of.author.get_recipient_for_visibility(content.visibility))
         logger.debug("send_share - sending to recipients: %s", recipients)
         handle_send(entity, content.author.federable, recipients, payload_logger=get_outbound_payload_logger())
+        target_content = content.share_of
+        if target_content.replies_fid:
+            queue = django_rq.get_queue('low')
+            content_id = target_content.id if target_content.content_type == ContentType.CONTENT else target_content.root_parent_id
+            if django_rq.get_scheduler(queue=queue).enqueue_in(timedelta(seconds=90),
+                    process_replies, content_id, shared_by_id=content.id):
+                logger.info("send_share - queued process_replies job for content id %s", content_id)
+            else:
+                logger.warn("send_share - failed to enqueue process_replies job for content id %s", content_id)
     else:
         logger.warning("send_share - No entity for %s", content)
 
@@ -233,7 +246,7 @@ def send_content_retraction(content, author_id):
     author = Profile.objects.get(id=author_id)
     entity = make_federable_retraction(content, author)
     if entity:
-        if settings.DEBUG:
+        if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
             # Don't send in development mode
             return
         if content.visibility == Visibility.PUBLIC:
@@ -265,7 +278,7 @@ def send_profile_retraction(profile):
         return
     entity = make_federable_retraction(profile)
     if entity:
-        if settings.DEBUG:
+        if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
             # Don't send in development mode
             return
         recipients = _get_remote_followers(profile, profile.visibility)
@@ -294,7 +307,7 @@ def forward_entity(entity, target_content_id):
     except Content.DoesNotExist:
         logger.warning("forward_entity - No content found with uuid %s", entity.id)
         return
-    if settings.DEBUG:
+    if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
         # Don't send in development mode
         return
     if target_content.visibility == Visibility.PUBLIC:
@@ -327,11 +340,12 @@ def send_follow_change(profile_id, followed_id, follow):
     except Profile.DoesNotExist:
         logger.warning("send_follow_change - No remote profile %s found to send follow for", followed_id)
         return
-    if settings.DEBUG:
+    if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
         # Don't send in development mode
         return
+    activity = Activity.objects.filter(profile=profile, object_id=remote_profile.id, type=ActivityType.FOLLOW).last()
     entity = base.Follow(
-        activity_id=f'{profile.fid}#follow-{uuid4()}',
+        activity_id=getattr(activity, 'fid', f'{profile.fid}#follow-{uuid4()}'),
         actor_id=profile.fid,
         target_id=remote_profile.fid,
         following=follow,
@@ -343,7 +357,7 @@ def send_follow_change(profile_id, followed_id, follow):
     logger.debug("send_follow_change - sending to recipients: %s", recipients)
     handle_send(entity, profile.federable, recipients, payload_logger=get_outbound_payload_logger())
     # Also trigger a profile send
-    send_profile(profile_id, recipients=recipients)
+    if follow: send_profile(profile_id, recipients=recipients)
 
 
 def send_profile(profile_id, recipients=None):
@@ -361,7 +375,9 @@ def send_profile(profile_id, recipients=None):
     if not entity:
         logger.warning("send_profile - No entity for %s", profile)
         return
-    if settings.DEBUG:
+    if hasattr(entity, 'times'):
+        entity.times['edited'] = (datetime.now(timezone.utc)-profile.modified)< timedelta(seconds=60)
+    if settings.DEBUG and settings.SOCIALHOME_DOMAIN.startswith('127.0'):
         # Don't send in development mode
         return
     if not recipients:
@@ -371,6 +387,10 @@ def send_profile(profile_id, recipients=None):
         else:
             recipients = []
         recipients.extend(_get_remote_followers(profile, profile.visibility))
+        to = NAMESPACE_PUBLIC
+    else:
+        to = recipients
+    if hasattr(entity, 'to'): entity.to = to
 
     logger.debug("send_profile - sending to recipients: %s", recipients)
     handle_send(entity, profile.federable, recipients, payload_logger=get_outbound_payload_logger())