+++ socialhome/federate/utils/tasks.py
@@ -1,9 +1,12 @@
-import logging
 import datetime as dt
+import logging
 from typing import Optional, List, Any
 
 import django_rq
+from django.conf import settings
+
 from federation.entities import base
+from federation.entities.activitypub.models import extract_replies
 from federation.fetchers import retrieve_remote_profile, retrieve_remote_content
 
 from socialhome.content.enums import ContentType
@@ -117,6 +120,8 @@ def process_entity_post(entity: Any, profile: Profile):
         "remote_created": safe_make_aware(entity.created_at, "UTC"),
         "service_label": safe_text(entity.provider_display_name) or "",
     }
+    if isinstance(getattr(entity, 'replies', None), base.Collection):
+        values["replies_fid"] = entity.replies.id
     extra_lookups = {}
     if getattr(entity, "guid", None):
         values["guid"] = safe_text(entity.guid)
@@ -125,17 +130,20 @@ def process_entity_post(entity: Any, profile: Profile):
     _process_mentions(content, entity)
     if created:
         logger.info("Saved Content: %s", content)
-        if hasattr(entity, '_replies'):
+        if content.replies_fid:
             queue = django_rq.get_queue("low")
-            if django_rq.get_scheduler(queue=queue).enqueue_in(dt.timedelta(seconds=90), process_replies, entity):
-                logger.info("process_replies - queued job for entity %s", entity.id)
+            if django_rq.get_scheduler(queue=queue).enqueue_in(dt.timedelta(seconds=120),
+                    process_replies, content.id):
+                logger.info("process_entity_post - queued process_replies job for entity %s", entity.id)
             else:
-                logger.warning("process_replies - failed to enqueue job for entity %s", entity.id)
+                logger.warning("process_entity_post - failed to enqueue process_replies job for entity %s", entity.id)
     else:
         logger.info("Updated Content: %s", content)
     if content.visibility == Visibility.LIMITED:
+        # add author to visibilities for local reply retractions to work.
+        receivers = [profile]
         if entity._receivers:
-            receivers = get_profiles_from_receivers(entity._receivers)
+            receivers.extend(get_profiles_from_receivers(entity._receivers))
             if len(receivers):
                 content.limited_visibilities.set(receivers)
                 logger.info("Added visibility to Post %s to %s", content.fid, receivers)
@@ -199,6 +207,8 @@ def process_entity_comment(entity: Any, profile: Profile):
         "parent": parent,
         "root_parent": root_parent,
     }
+    if isinstance(getattr(entity, 'replies', None), base.Collection):
+        values["replies_fid"] = entity.replies.id
     extra_lookups = {}
     if getattr(entity, "guid", None):
         values["guid"] = safe_text(entity.guid)
@@ -207,12 +217,6 @@ def process_entity_comment(entity: Any, profile: Profile):
     _process_mentions(content, entity)
     if created:
         logger.info("Saved Content from comment entity: %s", content)
-        if hasattr(entity, '_replies'):
-            queue = django_rq.get_queue("low")
-            if django_rq.get_scheduler(queue=queue).enqueue_in(dt.timedelta(seconds=90), process_replies, entity):
-                logger.info("process_replies - queued job for entity %s", entity.id)
-            else:
-                logger.warning("process_replies - failed to enqueue job for entity %s", entity.id)
     else:
         logger.info("Updated Content from comment entity: %s", content)
 
@@ -273,7 +277,7 @@ def _process_mentions(content, entity):
     Link mentioned profiles to the content.
     """
     fids = set(entity._mentions)
-    existing_fids = set(content.mentions.values_list('fid', flat=True))
+    existing_fids = set(content.mentions.values_list('finger', flat=True))
     to_remove = existing_fids.difference(fids)
     to_add = fids.difference(existing_fids)
     for fid in to_remove:
@@ -382,6 +386,14 @@ def process_entity_share(entity, profile):
     _process_mentions(content, entity)
     if created:
         logger.info("Saved share: %s", content)
+        if target_content.replies_fid:
+            queue = django_rq.get_queue('low')
+            content_id = target_content.id if target_content.content_type == ContentType.CONTENT else target_content.root_parent_id
+            if django_rq.get_scheduler(queue=queue).enqueue_in(dt.timedelta(seconds=90),
+                    process_replies, content_id, shared_by_id=content.id):
+                logger.info("process_entity_share - queued process_replies job for content id %s", content_id)
+            else:
+                logger.warn("process_entity_share - failed to enqueue process_replies job for content id %s", content_id)
     else:
         logger.info("Updated share: %s", content)
     # TODO: send participation to the share from the author, if local
@@ -393,35 +405,65 @@ def process_entity_share(entity, profile):
         queue.enqueue(forward_entity, entity, target_content.id)
 
 
-def process_replies(entity=None, fetch=False, delta=None):
+def process_replies(root_id, shared_by_id=None, delta=None):
     # Process Activitypub reply collection
-    if fetch:
-        # Refresh entity
-        entity = retrieve_remote_content(entity.id)
-        if not entity: return
+    try:
+        root = Content.objects.get(id=root_id)
+    except Content.DoesNotExist:
+        # Retracted?
+        return
+    # A job might have been scheduled from process_entity_shares
+    if shared_by_id:
+        if getattr(root.shares.last(), 'id', None) != shared_by_id:
+            logger.info("process_replies - job replaced by one scheduled from process_entity_share for most recent share of content id %s", root.fid)
+            return
+    elif root.shares_count > 0:
+        logger.info("process_replies - job replaced by one scheduled from process_entity_share for content id %s", root.fid)
+        return
 
+    def process_reply_collection(replies_fid):
+        coll = retrieve_remote_content(replies_fid, cache=False) # refresh reply collection
+        if isinstance(coll, base.Collection):
+            replies = extract_replies(getattr(coll, 'first', []))
+            for reply in replies:
+                reply_fid = getattr(reply, 'id', reply)
+                try:
+                    content = Content.objects.fed(reply_fid).get()
+                    if content.replies_fid: process_reply_collection(content.replies_fid)
+                except Content.DoesNotExist:
+                    # Try to fetch and process
+                    if isinstance(reply, base.Comment):
+                        remote_content = reply
+                    elif isinstance(reply, str):
+                        remote_content = retrieve_remote_content(reply_fid)
+                    else:
+                        continue
+                    if remote_content:
+                        logger.info(
+                            "process_replies - processing reply %s for entity %s",
+                            remote_content.id, remote_content.target_id)
+                        process_entities([remote_content])
+                        try:
+                            content = Content.objects.fed(reply_fid).get()
+                            if content.replies_fid: process_reply_collection(content.replies_fid)
+                        except Content.DoesNotExist:
+                            logger.warning(
+                                "process_replies - reply %s for entity %s could not be processed",
+                                remote_content.id, remote_content.target_id)
+                            continue
 
-    for reply in getattr(entity, '_replies', []):
-        try:
-            content = Content.objects.fed(reply).get()
-        except Content.DoesNotExist:
-            # Try to fetch and process
-            logger.debug(
-                "process_replies - fetching reply %s for entity %s", reply, entity.id)
-            remote_content = retrieve_remote_content(reply)
-            if remote_content:
-                process_entities([remote_content])
+    process_reply_collection(root.replies_fid)
 
     # Using a delta increasing by a factor of two, refresh
     # the replies up to 5 days after publication
+    if settings.DEBUG: return
     delta = delta * 2 if delta else dt.timedelta(minutes=15)
-    if hasattr(entity, '_replies'):
-        if delta < dt.timedelta(5):
-            queue = django_rq.get_queue("low")
-            if django_rq.get_scheduler(queue=queue).enqueue_in(delta, process_replies, entity, True, delta):
-                logger.info("process_replies - queued refresh job for entity %s", entity.id)
-            else:
-                logger.warning("process_replies - failed to enqueue refresh job for entity %s", entity.id)
+    if delta < dt.timedelta(3):
+        queue = django_rq.get_queue('low')
+        if django_rq.get_scheduler(queue=queue).enqueue_in(delta, process_replies, root_id, shared_by_id, delta):
+            logger.info("process_replies - queued refresh job for entity %s", root.fid)
+        else:
+            logger.warning("process_replies - failed to enqueue refresh job for entity %s", root.fid)
 
 
 def sender_key_fetcher(fid):